pyproject.toml
[tool.poetry]
name = "worker-api"
version = "0.1.0"
description = ""
authors = ["Michael <pozdinmisha666@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "^3.1.4"
schedule = "^1.2.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
aiogram = "^3.10.0"
asyncio = "^3.4.3"
aioschedule = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
